requirements.txt
click==7.1.2
Flask==2.1.0
Flask-MySQL==1.5.2
itsdangerous==1.1.0
Jinja2==3.0.2
MarkupSafe==2.1.0
mysql-connector-python==8.0.27
mysqlclient==2.1.0
Werkzeug==2.0.2
